requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
dnspython==2.2.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
py-cord==2.1.1
pymongo==4.1.1
python-dotenv==0.20.0
yarl==1.7.2
